requirements.txt
tensorflow==2.13.0
streamlit>=1.28.0
pillow>=10.0.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
numpy>=1.24.0